src/requirements.txt
aiohttp==1.1.1
async-timeout==1.1.0
chardet==2.3.0
mongoengine==0.10.6
multidict==2.1.2
pymongo==3.3.1
PyYAML==3.12
yarl==0.5.3
